requirements.txt

yarl==1.18.3
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.8.3
aiomysql==0.2.0
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
bcrypt==4.2.1
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==43.0.3
cycler==0.12.1
defusedxml==0.7.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
faiss-cpu==1.9.0.post1
fastadmin==0.2.16
fastapi==0.115.5
filelock==3.16.1
filetype==1.2.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
gauthuserinfo==1.0.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
gunicorn==23.0.0
h11==0.14.0
httplib2==0.22.0
huggingface-hub==0.27.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numpy==2.2.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
passlib==1.7.4
pdfkit==1.0.0
pillow==11.0.0
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
pydicom==3.0.1
PyJWT==2.10.0
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-multipart==0.0.17
PyYAML==6.0.2
razorpay==1.4.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.49
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.14.1
sentence-transformers==3.3.1
six==1.17.0
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.36
starlette==0.41.3
supervision==0.25.0
sympy==1.13.1
threadpoolctl==3.5.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
twilio==9.4.2
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uuid==1.30
uvicorn==0.32.1
websockets==14.1
WTForms==3.1.2
yarl==1.18.3
